fixed some moments
