+++ src/pages/index.tsx
@@ -1,8 +1,9 @@
 import { FC } from 'react';
 import { Navigate, Route, Routes } from 'react-router-dom';
-import { privateRoutes, publicRoutes, routesEnum } from './types';
-import PrivateRoute from './PrivateRoute';
+import { AuthEnum, routesEnum } from './types';
+import PrivateRoute from './private-route';
 import { OfferType } from '../shared/interface';
+import { privateRoutes, publicRoutes } from './routes';
 
 interface IAppRouterProps {
   offersMockData: OfferType[];
@@ -12,7 +13,7 @@ const AppRouter: FC<IAppRouterProps> = (props: IAppRouterProps) => (
     {publicRoutes.map((myRoute) => <Route key={myRoute.path} element={<myRoute.Component {...props}/>} path={myRoute.path} />)}
     {privateRoutes.map((myRoute) =>(
       <Route key={myRoute.path} path={myRoute.path}
-        element={<PrivateRoute key={myRoute.path}>{<myRoute.Component {...props}/>}</PrivateRoute>}
+        element={<PrivateRoute authState={AuthEnum.AUTHENTICATED} key={myRoute.path}>{<myRoute.Component {...props}/>}</PrivateRoute>}
       />
     )
     )}
