+++ src/pages/not-found-page.tsx
@@ -1,9 +1,10 @@
 import { Link } from 'react-router-dom';
+import { routesEnum } from './types';
 
 const NotFoundPage = () => (
   <div>
     <h1>404 Not Found</h1>
-    <Link to='/'>Return to main page</Link>
+    <Link to={routesEnum.MAIN}>Return to main page</Link>
   </div>
 );
