+++ src/pages/types.ts
@@ -1,12 +1,6 @@
 import { ElementType } from 'react';
-import FavoritesPage from './FavoritesPage';
-import LoginPage from './LoginPage';
-import MainPage from './MainPage';
-import NotFoundPage from './NotFoundPage';
-import OfferPage from './OfferPage';
 
-
-interface IRoute {
+export interface IRoute {
   path: routesEnum;
   Component: ElementType;
 }
@@ -17,29 +11,10 @@ export enum routesEnum {
   OFFER = '/offer/:id',
   NOT_FOUND = '/notFound'
 }
-export const publicRoutes: IRoute[] = [
-  {
-    path: routesEnum.LOGIN,
-    Component: LoginPage,
-  },
-  {
-    path: routesEnum.MAIN,
-    Component: MainPage,
-  },
-  {
-    path: routesEnum.OFFER,
-    Component: OfferPage,
-  },
-  {
-    path: routesEnum.NOT_FOUND,
-    Component: NotFoundPage,
-  }
-];
 
-export const privateRoutes: IRoute[] = [
-  {
-    path: routesEnum.FAVORITES,
-    Component: FavoritesPage,
-  }
-];
+export enum AuthEnum {
+  AUTHENTICATED,
+  NO_AUTHENTICATED,
+  UNKNOWN
+}
